requirements-dev.txt
-r requirements.txt

mock
pytest
Flask-Testing
ipdb
ipython
